device: Move class-c tests into their own file
